rename according to flutter guideline
